refactor: return controller in components dir
